Split footer into a new file
