+++ web/src/pages/Home.tsx
@@ -4,6 +4,7 @@ import Header from "../components/Header";
 import LogoImage from "../assets/logo.svg";
 import { LinkButton } from "../components/LinkButton";
 import BuyMeACoffee from "../components/BuyMeACoffee";
+import Footer from "../components/Footer";
 
 const Hero: Component = () => {
   return (
@@ -71,16 +72,6 @@ const Features: Component = () => {
     </div>
   );
 };
-
-const Footer: Component = () => (
-  <footer class="flex flex-col sm:flex-row items-center justify-center px-8 py-5 text-center">
-    <span class="text-slate-400 font-normal">
-      MIT Licensed | Copyright © 2021-present Haneen Mahdin
-    </span>
-
-    <BuyMeACoffee />
-  </footer>
-);
 
 const Home: Component = () => {
   return (
